refactor: move frontend types in app folder
