+++ src/app/_components/Navbar.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import Link from "next/link";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import type { NavbarLink, NavbarProps } from "~/types";
+import type { NavbarLink, NavbarProps } from "~/app/types";
 
 export function Navbar(props: NavbarProps) {
   const router = useRouter();
