+++ src/app/types/navigation.ts
