+++ src/constants/index.tsx
@@ -3,7 +3,7 @@ import { dashboardNavLinks } from "./navbarLinks";
 import { BacklogEntryDummyData } from "./backlogEntryDummyData";
 import { gameSearchbarDummyResults } from "~/constants/gameSearchbarDummyResults";
 
-export { type NavbarLink } from "~/types";
+export { type NavbarLink } from "~/app/types";
 export {
   landingPageNavLinks,
   dashboardNavLinks,
